backend/requirements.txt
# Core Flask Framework
Flask==3.0.0
flask-cors==4.0.0

# Environment Variables
python-dotenv==1.0.0

# File Processing
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.1.0

# AWS SDK
boto3==1.29.7

# Snowflake Connector
snowflake-connector-python==3.7.0

# Testing (optional but recommended)
pytest==7.4.3
